# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.ServerGroupResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Asia Pacific
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  model.ShadowsocksServerResponse:
    properties:
      cipher:
        example: aes-256-gcm
        type: string
      created_at:
        example: 1640995200
        type: integer
      excludes:
        example: 192.168.0.0/16
        type: string
      group_id:
        example: 1
        type: integer
      host:
        example: us01.example.com
        type: string
      id:
        example: 1
        type: integer
      ips:
        example: 0.0.0.0/0
        type: string
      name:
        example: US-01
        type: string
      obfs:
        example: tls
        type: string
      obfs_settings:
        example: obfs=tls
        type: string
      parent_id:
        example: 1
        type: integer
      port:
        example: 443
        type: integer
      rate:
        example: 1
        type: number
      route_id:
        example: route-1
        type: string
      server_group:
        allOf:
        - $ref: '#/definitions/model.ServerGroupResponse'
        description: Relationship data
      server_port:
        example: 8388
        type: integer
      show:
        example: 1
        type: integer
      sort:
        example: 1
        type: integer
      tags:
        example: premium,fast
        type: string
      updated_at:
        example: 1640995200
        type: integer
    type: object
...
  response.PaginatedResponse:
    description: Paginated response format
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          type: object
        type: array
      limit:
        example: 10
        type: integer
      message:
        example: success
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  response.StandardResponse:
    description: Standard API response format
    properties:
      code:
        description: Response code (0 for success, non-zero for errors)
        example: 0
        type: integer
      data:
        description: Response data (optional)
        type: object
      message:
        description: Response message
        example: success
        type: string
    type: object
...
  service.CreateShadowsocksServerRequest:
    properties:
      cipher:
        maxLength: 255
        type: string
      excludes:
        type: string
      group_id:
        type: integer
      host:
        maxLength: 255
        type: string
      ips:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      obfs:
        maxLength: 11
        type: string
      obfs_settings:
        maxLength: 255
        type: string
      parent_id:
        type: integer
      port:
        maximum: 65535
        minimum: 1
        type: integer
      rate:
        minimum: 0.1
        type: number
      route_id:
        maxLength: 255
        type: string
      server_port:
        maximum: 65535
        minimum: 1
        type: integer
      show:
        maximum: 1
        minimum: 0
        type: integer
      sort:
        type: integer
      tags:
        maxLength: 255
        type: string
    required:
    - cipher
    - group_id
    - host
    - name
    - port
    - rate
    - server_port
    type: object
...
  service.UpdateShadowsocksServerRequest:
    properties:
      cipher:
        maxLength: 255
        type: string
      excludes:
        type: string
      group_id:
        type: integer
      host:
        maxLength: 255
        type: string
      ips:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      obfs:
        maxLength: 11
        type: string
      obfs_settings:
        maxLength: 255
        type: string
      parent_id:
        type: integer
      port:
        maximum: 65535
        minimum: 1
        type: integer
      rate:
        minimum: 0.1
        type: number
      route_id:
        maxLength: 255
        type: string
      server_port:
        maximum: 65535
        minimum: 1
        type: integer
      show:
        maximum: 1
        minimum: 0
        type: integer
      sort:
        type: integer
      tags:
        maxLength: 255
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive service management platform with subscription-based
    billing, user management, and server administration. Features include OAuth2 authentication,
    traffic subscription management, multi-gateway payments, referral programs, and
    customer support system.
  title: Linke API
  version: "1.0"
paths:
...
  /admin/shadowsocks-servers:
    get:
      consumes:
      - application/json
      description: Get shadowsocks servers with optional filters (Admin only)
      parameters:
      - description: Group ID filter
        in: query
        name: group_id
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Show filter
        in: query
        name: is_show
        type: boolean
      - description: Online filter
        in: query
        name: is_online
        type: boolean
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get Shadowsocks Servers
      tags:
      - Admin-ShadowsocksServers
    post:
      consumes:
      - application/json
      description: Create a new shadowsocks server (Admin only)
      parameters:
      - description: Create shadowsocks server request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateShadowsocksServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ShadowsocksServerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create Shadowsocks Server
      tags:
      - Admin-ShadowsocksServers
  /admin/shadowsocks-servers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a shadowsocks server (Admin only)
      parameters:
      - description: Shadowsocks Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete Shadowsocks Server
      tags:
      - Admin-ShadowsocksServers
    get:
      consumes:
      - application/json
      description: Get a shadowsocks server by ID (Admin only)
      parameters:
      - description: Shadowsocks Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ShadowsocksServerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get Shadowsocks Server by ID
      tags:
      - Admin-ShadowsocksServers
    patch:
      consumes:
      - application/json
      description: Partially update a shadowsocks server using PATCH method (Admin
        only)
      parameters:
      - description: Shadowsocks Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Partial shadowsocks server update data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ShadowsocksServerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Partially Update Shadowsocks Server
      tags:
      - Admin-ShadowsocksServers
    put:
      consumes:
      - application/json
      description: Update a shadowsocks server (Admin only)
      parameters:
      - description: Shadowsocks Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update shadowsocks server request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateShadowsocksServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ShadowsocksServerResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update Shadowsocks Server
      tags:
      - Admin-ShadowsocksServers
...
  /user/shadowsocks-servers:
    get:
      consumes:
      - application/json
      description: Get available shadowsocks servers for current user based on subscription
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ShadowsocksServerResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get Available Shadowsocks Servers
      tags:
      - User-ShadowsocksServers
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"